rewrite viewer to make the usage more explicit
